+++ Assets/Prefabs/UI/Popups.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3b084d6de581348d3afac856a54ad2d9
+guid: b06fffe2b3c214d7ab2c0dc47dbcced6
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
