+++ Assets/Prefabs/UI/Popups/MainMenu.prefab
@@ -925,6 +925,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0b0da0be2c5b34d7f92c02269ef32230, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  _playButton: {fileID: 9116505314255488959}
 --- !u!1 &7002228691092847502
 GameObject:
   m_ObjectHideFlags: 0
