--- Assets/Prefabs/UI/StartSplashScreen.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3361be6f8b656bb4eb7799db1d99d708
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
